add: ajout du lien entre les catégories et les médias

+++ poorterest/database/migrations/0001_01_01_000000_create_medias_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('description');
             $table->string('size');
             $table->string('category');
-            $table->int('userId');
+            $table->unsignedBigInteger('userId');
             $table->string('status');
             $table->timestamps();
         });
